+++ src/test/java/org/jabref/logic/integrity/AmpersandCheckerTest.java
@@ -58,4 +58,18 @@ public class AmpersandCheckerTest {
                 Arguments.of("Found 4 unescaped '&'", StandardField.AFTERWORD, "May the force be with you & live long \\\\& prosper \\& to infinity \\\\\\& beyond & assemble \\\\\\\\& excelsior!")
         );
     }
+
+    static Stream<Arguments> entryWithVerabitmFieldsNotCausingMessages() {
+        return Stream.of(
+                Arguments.of(StandardField.FILE, "one & another.pdf"),
+                Arguments.of(StandardField.URL, "https://example.org?key=value&key2=value2")
+        );
+    }
+
+    @ParameterizedTest
+    @MethodSource
+    void entryWithVerabitmFieldsNotCausingMessages(Field field, String value) {
+        entry.setField(field, value);
+        assertEquals(List.of(), checker.check(entry));
+    }
 }